third_party/blink/web_tests/fast/frames/detached-shadow-frame-expected.txt

Should not be able to create an iframe with a loaded contentDocument that is not in the document tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS innerFrame.contentDocument is null
PASS successfullyParsed is true

TEST COMPLETE
